finished abstracting the code from flagged_clients into the flag_model and ensured tests still work. Need to remove comments from the flagged_clients model
